% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-post-proc.R
\name{add_aggregated_markdown_to_eddy_postproc}
\alias{add_aggregated_markdown_to_eddy_postproc}
\title{Add aggregated markdown to EDDY post_proc}
\usage{
add_aggregated_markdown_to_eddy_postproc(eddy_postproc)
}
\arguments{
\item{eddy_postproc}{list: ...}
}
\value{
eddy_postproc list: ...
}
\description{
Add aggregated markdown to EDDY post_proc
}
