% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-graph.R
\name{add_mediators_to_DDN_graph}
\alias{add_mediators_to_DDN_graph}
\title{Add mediators to DDN graph}
\usage{
add_mediators_to_DDN_graph(ddn_graph, mediator_tbl)
}
\arguments{
\item{ddn_graph}{a ggraph object of DDN table}

\item{mediator_tbl}{table made of name (mediator) and
additional columns to indicate mediator types}
}
\value{
ddn_graph amended with mediators
}
\description{
Add mediators to DDN graph
}
