% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-post-proc.R
\name{add_summary_table_markdown_to_eddy_postproc}
\alias{add_summary_table_markdown_to_eddy_postproc}
\title{Add summary table in markdown format}
\usage{
add_summary_table_markdown_to_eddy_postproc(eddy_postproc)
}
\arguments{
\item{eddy_postproc}{list: ...}
}
\value{
eddy_postproc list: ...
}
\description{
Add summary table in markdown format
}
