% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-glassonet.R
\name{aggregate_DDN_table}
\alias{aggregate_DDN_table}
\title{Combine all DDNs into a single table}
\usage{
aggregate_DDN_table(glasso_DDN_list)
}
\arguments{
\item{glasso_DDN_list}{list of glass DDNs}
}
\value{
data frame, DDN table
}
\description{
Combine all DDNs into a single table
}
