% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-graph.R
\name{as_DDN_graph}
\alias{as_DDN_graph}
\title{Create DDN graph from DDN table}
\usage{
as_DDN_graph(ddn_tbl, mediators = NULL)
}
\arguments{
\item{ddn_tbl}{DDN table.  DDN table should have \code{node_src} and \code{node_dst}, and
Pathway columns.}

\item{mediators}{mediators.  See \code{\link{add_mediators_to_DDN_graph}}.}
}
\value{
DDN graph which is a ggraph object of DDN table
}
\description{
If there is no condition in DDN table, its considered a single-conditioned
  network.  It also adds mediators to the graph if given.
}
\seealso{
[add_mediators_to_DDN_graph()]
}
