% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-glassonet.R
\name{calc_DDN_score}
\alias{calc_DDN_score}
\title{Calculate DDN score for ddn_tbl}
\usage{
calc_DDN_score(ddn_tbl)
}
\arguments{
\item{ddn_tbl}{object returned from \code{\link{to_edge_list}}}
}
\value{
list of prob, rewiring, and other parameters computed from ddn_tbl
}
\description{
Calculate DDN score for ddn_tbl
}
