% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-glassonet.R
\name{calc_glasso_DDN_score}
\alias{calc_glasso_DDN_score}
\title{Calculate DDN score for glasso_ddn}
\usage{
calc_glasso_DDN_score(glasso_ddn)
}
\arguments{
\item{glasso_ddn}{object returned from \code{\link{to_glasso_DDN}}}
}
\value{
output of \code{\link{calc_DDN_score}}
}
\description{
Calculate DDN score for glasso_ddn
}
