% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-graph.R
\name{compute_DDN_mediator_specificity}
\alias{compute_DDN_mediator_specificity}
\title{Compute specificity mediators for DDN}
\usage{
compute_DDN_mediator_specificity(ddn, p_val.cutoff = 0.05)
}
\description{
Compute specificity mediators for DDN
}
