% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-graph.R
\name{compute_DDN_mediators}
\alias{compute_DDN_mediators}
\title{Compute mediators for DDN}
\usage{
compute_DDN_mediators(
  ddn,
  mediator.type = "both",
  essentiality_cutoff = 0.05,
  specificity_cutoff = 0.05
)
}
\description{
Compute mediators for DDN
}
