% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-graph.R
\name{compute_DDN_mediators_by_pathway}
\alias{compute_DDN_mediators_by_pathway}
\title{Compute mediators for DDN by pathway}
\usage{
compute_DDN_mediators_by_pathway(ddn, ...)
}
\description{
Compute mediators for DDN by pathway
}
