% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-glassonet.R
\name{cv_glasso}
\alias{cv_glasso}
\title{Train GLASSO over cross-validation}
\usage{
cv_glasso(
  X,
  name = "cv_glasso",
  K = 5,
  iter_cv = 10,
  rholist = NA,
  thr = 1e-04,
  maxit = 10000,
  crit.cv = c("BIC", "AIC", "loglik"),
  approx = FALSE,
  penalize_diagonal = TRUE,
  rho_min_ratio = 0.01,
  trace = 0,
  progress_bar = TRUE,
  cores = 1
)
}
\arguments{
\item{X}{data}

\item{name}{name for GLASSO}

\item{K}{k for k-fold cross validation}

\item{iter_cv}{iteration of CV}

\item{rholist}{list of rho values.  If NULL, they will be automatically created.}

\item{thr}{threshold for convergence in \code{\link{glasso}}}

\item{maxit}{passed to \code{\link{glasso}}}

\item{crit.cv}{cross validation criterion (\code{BIC}, \code{AIC}, or \code{loglik}). Defaults to \code{BIC}.}

\item{approx}{passed to \code{\link{glasso}}}

\item{rho_min_ratio}{used to set minimum rho value}

\item{trace}{passed to \code{\link{glasso}}}

\item{cores}{number of cores for parallel processes. Use all available cores if 0.}

\item{penalized_diagonal}{passed to \code{\link{glasso}}}
}
\description{
Train GLASSO over cross-validation
}
