% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-table.R
\name{db_name_cleanup}
\alias{db_name_cleanup}
\alias{reactome_cleanup}
\title{Clean up Pathway DB name}
\usage{
db_name_cleanup(pathway_name, db_name)

reactome_cleanup(pathway_name)
}
\arguments{
\item{pathway_name}{name for pathway}

\item{db_name}{prefix indicating pathway database, like "REACTOME"}
}
\value{
new pathway name without prefix and postfix
}
\description{
Remove some prefix, like "REACTOME", specified by db_name,
  and postfix, like "_PATHWAY" in pathway_name
}
