% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-file-io.R
\name{ddn_to_json}
\alias{ddn_to_json}
\title{Convert DDN table to JSON string and save it to a file}
\usage{
ddn_to_json(ddn_tbl, filepath = NULL, return_JSON = FALSE)
}
\arguments{
\item{ddn_tbl}{data frame: DDN table}

\item{filepath}{string: path to a file or folder.
If not given, save to a current folder with pathway name as file name.}

\item{return_JSON}{logical (default FALSE): if true, JSON string is returned}
}
\description{
Convert DDN table to JSON string and save it to a file
}
