% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-table.R
\name{expand_mediators}
\alias{expand_mediators}
\title{Expand a pathway line with multiple mediators}
\usage{
expand_mediators(p)
}
\arguments{
\item{p}{a pathway line in summary table}
}
\value{
expanded summary table
}
\description{
When there are multiple mediators in a pathway,
  the pathway line is replicated in such a way
  that each line contains only one mediator and its type.
}
