% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-table.R
\name{flatten_DDN_mediators}
\alias{flatten_DDN_mediators}
\title{Flatten expanded summary/mediator table}
\usage{
flatten_DDN_mediators(mediator_tbl)
}
\description{
Flatten expanded summary/mediator table
}
