% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-table.R
\name{get_mediators}
\alias{get_mediators}
\title{Retrieve mediators as an array}
\usage{
get_mediators(p, type = "both")
}
\arguments{
\item{p}{a pathway line in summary table}

\item{type}{which type of mediators to retrieve}
}
\value{
array of mediators
}
\description{
Mediators are stored as a string.  This function parse them
  and return them in an array.  One can retried either specificity
  mediators or essentiality mediators or both simultaneously.
}
