% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-graph.R
\name{get_node_set}
\alias{get_node_set}
\title{Retrieve all nodes in table of DDNs corresponding to pathway_name}
\usage{
get_node_set(ddns_tbl, pathway_name)
}
\arguments{
\item{ddns_tbl}{a table of DDNs}

\item{pathway_name}{Name of pathway to retrieve the nodes for}
}
\value{
array of nodes

Obsolete????  Not used in anywhere else.
}
\description{
DDN table should have \code{node_src} and \code{node_dst}, and
  Pathway columns.
}
