% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-table.R
\name{parse_mediators}
\alias{parse_mediators}
\title{Parse mediator string}
\usage{
parse_mediators(s, sep = "\\\\s+")
}
\arguments{
\item{s}{string made of mediators}

\item{sep}{separator between mediators, DEFAULT is space/white character.}
}
\value{
a list of mediators or a string if there is only one mediator
}
\description{
In mediator columns (essentiality or specificity),
  mediators are separated by \code{sep},
  and convert them into a list of mediators or
  a string if there is only one mediator.
}
