% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-graph.R
\name{plot_DDN_graph}
\alias{plot_DDN_graph}
\title{Plot DDN graph}
\usage{
plot_DDN_graph(
  ddn_graph,
  group2node = NULL,
  graph_layout = "fr",
  show_node_label = FALSE,
  show_mediator_label = TRUE
)
}
\arguments{
\item{ddn_graph}{see \code{\link{as_DDN_graph}}}

\item{group2node}{two column table where the first column is group/pathway names, and the second column is node/gene names.}

\item{graph_layout}{Layout algorithm name, for example, "fr".  See \code{\link{layout_nicely}}.}

\item{show_node_label}{if TRUE, node/gene names are shown}

\item{show_mediator_label}{if TRUE, mediator names are shown}
}
\value{
ggraph/ggplot2 object
}
\description{
Plot DDN graph
}
