% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-graph.R
\name{plot_DDN_tbl_graph}
\alias{plot_DDN_tbl_graph}
\title{Plot DDN graph from DDN table}
\usage{
plot_DDN_tbl_graph(
  ddn_tbl,
  show_mediators = "both",
  show_pathway_group = TRUE,
  ...
)
}
\arguments{
\item{ddn_tbl}{see \code{\link{as_DDN_graph}}}

\item{show_mediators}{If show_mediators is one of 'none' (no mediator), 'specificity', 'essentiality' and 'both',
mediators will be computed and added to DDN graph and shown.  One can also provide a list of mediators.}

\item{show_pathway_group}{If TRUE, pathway names will be printed around the genes belonging to the same pathway.}

\item{...}{Additional parameters passed to \code{\link{plot_DDN_graph}}}
}
\value{
ggraph/ggplot2 object
}
\description{
Plot DDN graph from DDN table
}
