% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-post-proc.R
\name{post_proc_EDDY}
\alias{post_proc_EDDY}
\title{Post process EDDY-GPU outputs}
\usage{
post_proc_EDDY(
  eddy_run_dir,
  create_ddn_graph = FALSE,
  mapping_samples_to_condition = NULL,
  db_name_prefix = NA
)
}
\arguments{
\item{eddy_run_dir}{...}

\item{create_ddn_graph}{...}

\item{mapping_samples_to_condition}{...}

\item{db_name_prefix}{...}
}
\value{
...
}
\description{
Post process EDDY-GPU outputs
}
