% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-post-proc.R
\name{post_proc_EDDY_DDNs}
\alias{post_proc_EDDY_DDNs}
\alias{post_proc_EDDY_folder}
\alias{post_proc_EDDY_glasso}
\title{Post process EDDY DDNs and summary table}
\usage{
post_proc_EDDY_DDNs(
  DDNs,
  summary_tbl,
  create_ddn_graph = FALSE,
  mapping_conditions = NA,
  db_name_prefix = NA
)

post_proc_EDDY_folder(
  eddy_run_dir,
  create_ddn_graph = FALSE,
  mapping_conditions = NA,
  db_name_prefix = NA
)

post_proc_EDDY_glasso(
  glasso_DDN_list,
  create_ddn_graph = FALSE,
  mapping_conditions = NA,
  db_name_prefix = NA
)
}
\arguments{
\item{DDNs}{data.frame (DDN table)}

\item{summary_tbl}{data.frame}

\item{create_ddn_graph}{logical (default is FALSE)}

\item{db_name_prefix}{string, like REACTOME}

\item{eddy_run_dir}{string (path)}

\item{glasso_DDN_list}{list of glasso DDNs}

\item{mapping_samples_to_condition}{named vector}
}
\value{
list of 1) summary tabla, 2) DDNs (data frame),
        3) a list of DDN graphs,
        4) aggregated DDN graph (full), and
        5) aggregated DDN graph (p < 0.05)
}
\description{
Post process EDDY DDNs and summary table

Post process EDDY-GPU outputs

Post process EDDY-glasso-DDN outputs
}
