% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-post-proc.R
\name{post_proc_EDDY_folder}
\alias{post_proc_EDDY_folder}
\title{Post process EDDY-GPU outputs}
\usage{
post_proc_EDDY_folder(
  eddy_run_dir,
  create_ddn_graph = FALSE,
  mapping_samples_to_condition = NULL,
  db_name_prefix = NULL
)
}
\arguments{
\item{eddy_run_dir}{string (path)}

\item{create_ddn_graph}{logical (default is FALSE)}

\item{mapping_samples_to_condition}{named vector}

\item{db_name_prefix}{string, like REACTOME}
}
\value{
list of 1) summary tabla, 2) DDNs (data frame), 3) a list of DDN graphs,
        4) aggregated DDN graph (full), and 5) aggregated DDN graph (p < 0.05)
}
\description{
Post process EDDY-GPU outputs
}
