% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-file-io.R
\name{read_DDN_EdgeList}
\alias{read_DDN_EdgeList}
\title{Read DDN edgelist}
\usage{
read_DDN_EdgeList(ddn_filepath)
}
\arguments{
\item{ddn_filepath}{a path to *_Edgelist.txt}
}
\value{
\code{tibble} table of DDN edge list.
  Table has 'node_src', 'node_dst', 'condition', 'prior', and 'pathway' columns.
}
\description{
Read DDN (*_EdgeList.txt) file.
}
