% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-file-io.R
\name{read_DDNs}
\alias{read_DDNs}
\title{Read all DDNs in a folder}
\usage{
read_DDNs(path)
}
\arguments{
\item{path}{path to a folder with all DDNs (*_EdgeList.txt files)}
}
\value{
Combined (\code{bind_rows}) DDN table
}
\description{
Read all DDNs in a folder
}
