% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-file-io.R
\name{read_results_txt}
\alias{read_results_txt}
\title{Read 'results.txt'}
\usage{
read_results_txt(run.dir)
}
\arguments{
\item{run.dir}{a path to 'results.txt'}
}
\value{
\code{tibble} table
}
\description{
Use this function if EDDY results are not post-processed yet.
  This will be a recommended method in the future.
}
\details{
Read 'results.txt' which is a table with all statistics for each gene set run by EDDY.
  'results.txt' is space separated file and the format is a bit messy for those FAILED gene set.
  So, this function actually reads 'results.txt', remove those FAILED gene set/lines,
  reformat them nicely and save them into a tab-delimited file, 'results.tsv'.
}
