% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-file-io.R
\name{read_summary_table}
\alias{read_summary_table}
\title{Read summary table}
\usage{
read_summary_table(path, tag = NA)
}
\arguments{
\item{path}{path to summary table csv file}

\item{tag}{tag to indicate the study.  If not given,
the name of folder where summary table csv file is located.}
}
\value{
summary table
}
\description{
Use this function if EDDY results are already post-processed.
}
