% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-table.R
\name{to_DDN_summary}
\alias{to_DDN_summary}
\title{Generate summary table}
\usage{
to_DDN_summary(ddn_tbl)
}
\arguments{
\item{ddn_tbl}{data.frame}
}
\value{
summary table
}
\description{
Generate summary table
}
