% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-glassonet.R
\name{to_edge_list}
\alias{to_edge_list}
\title{Convert adjacent matrix into edge list}
\usage{
to_edge_list(
  adj_mat,
  diag = "off",
  symmetric = TRUE,
  signed = FALSE,
  simplify = TRUE
)
}
\arguments{
\item{adj_mat}{adjacent matrix}

\item{diag}{passed to \code{\link{shrink_adj_mat}}}

\item{symmetric}{force adj_mat to symmetric, also passed to \code{\link{shrink_adj_mat}}}

\item{signed}{if TRUE, edge will be marked either "inhibitory" or "synergistic".  If FALSE, all edges will be "connected".}

\item{simplify}{if TRUE, all disconnected edges will be removed}
}
\value{
edge list (data.frame)
}
\description{
Convert adjacent matrix into edge list
}
