% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-glassonet.R
\name{to_glasso_DDN}
\alias{to_glasso_DDN}
\title{convert list of glassonets to DDN}
\usage{
to_glasso_DDN(glnet_list)
}
\arguments{
\item{glnet_list}{a list of glassonets}
}
\value{
glasso_DDN list (a list of glassonets and ddn_tbl)
}
\description{
convert list of glassonets to DDN
}
