% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-glassonet.R
\name{to_glasso_DDN_summary}
\alias{to_glasso_DDN_summary}
\title{Generate summary table}
\usage{
to_glasso_DDN_summary(glasso_DDN_list, selected_names = NULL)
}
\arguments{
\item{glass_DDN_list}{list of glasso_DDNs}
}
\value{
summary table
}
\description{
Generate summary table
}
