% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-glassonet.R
\name{to_glasso_DDN_tbl}
\alias{to_glasso_DDN_tbl}
\title{Combine all DDNs into a single table}
\usage{
to_glasso_DDN_tbl(glasso_DDN_list)
}
\arguments{
\item{glasso_DDN_list}{list of glass DDNs}
}
\value{
data frame, DDN table
}
\description{
Combine all DDNs into a single table
}
