% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-glassonet.R
\name{to_glassonet}
\alias{to_glassonet}
\title{Convert glasso or cv_glasso into glassonet}
\usage{
to_glassonet(gls)
}
\arguments{
\item{gls}{glassor or cv_glasso}
}
\value{
glassonet (glasso, precision and partial correlation matrices, edge list)
}
\description{
Convert glasso or cv_glasso into glassonet
}
