% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-ddn-graph-json.R
\name{wrap_json_elem}
\alias{wrap_json_elem}
\title{Wrap JSON element with either "{ }" or "[ ]" pair}
\usage{
wrap_json_elem(x, wrapper = c("{", "["))
}
\description{
Wrap JSON element with either "{ }" or "[ ]" pair
}
