% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-post-proc.R
\name{write_eddy_post_processed}
\alias{write_eddy_post_processed}
\title{Write EDDY post-processed results}
\usage{
write_eddy_post_processed(
  eddy_post_processed,
  out_dir = ".",
  fig.width = 12,
  fig.height = 9
)
}
\arguments{
\item{eddy_post_processed}{...}

\item{out_dir}{...}

\item{fig.width}{...}

\item{fig.height}{...}
}
\value{
...s
}
\description{
Write EDDY post-processed results
}
