% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-post-proc.R
\name{write_eddy_postproc_csv}
\alias{write_eddy_postproc_csv}
\title{Write EDDY post-processed results into tables}
\usage{
write_eddy_postproc_csv(eddy_postproc, output_dir = ".")
}
\arguments{
\item{eddy_postproc}{...}

\item{output_dir}{...}
}
\value{
...s
}
\description{
Write EDDY post-processed results into tables
}
