% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-post-proc.R
\name{write_eddy_postproc_pdf}
\alias{write_eddy_postproc_pdf}
\title{Write EDDY post-processed results and DDNs into PDF}
\usage{
write_eddy_postproc_pdf(
  eddy_postproc,
  out_dir = ".",
  fig.width = 12,
  fig.height = 9
)
}
\arguments{
\item{eddy_postproc}{...}

\item{out_dir}{...}

\item{fig.width}{...}

\item{fig.height}{...}
}
\value{
...s
}
\description{
Write EDDY post-processed results and DDNs into PDF
}
