% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy-post-proc.R
\name{write_eddy_postproc_summary_table_to_markdown}
\alias{write_eddy_postproc_summary_table_to_markdown}
\title{Write summary table in markdown format}
\usage{
write_eddy_postproc_summary_table_to_markdown(eddy_postproc, output_dir)
}
\arguments{
\item{eddy_postproc}{list: ...}

\item{output_dir}{string: ...}
}
\value{
...s
}
\description{
Write summary table in markdown format
}
